Remove replace holder
