+++ cpp/src/gandiva/gdv_function_stubs.cc
@@ -37,7 +37,6 @@
 #include "gandiva/precompiled/types.h"
 #include "gandiva/random_generator_holder.h"
 #include "gandiva/regex_functions_holder.h"
-#include "gandiva/replace_holder.h"
 #include "gandiva/to_date_holder.h"
 
 /// Stub functions that can be accessed from LLVM or the pre-compiled library.
